Changed the forceSolver assumptions
Changed the force solver assumptions to allow it assume any multiple of forces are the same, also ended up making it a lot simpler to use. Works by adding lines to the system of equations which have a -1 and 1 pair.
Some minor changes to keep the test functions working as well.